requirements.txt
intelhex>=2.3.1
pyserial>=3.5
PyQt6>=6.0.0 